Apply factory pattern on Teltonika payloads
Apply the factory design pattern on Teltonika payloads to allow for easier and nicer look and feel of the code. Ideally, should be expanded to be used by all Telematics devices to be added to OpenRemote. Would be cool if we could use that interface to autocreate MQTT handlers for different device manufacturers.