Finish Teltonika Configuration Support.
Create TeltonikaConfigurationFactory to create and manage the instance

Edit TeltonikaConfiguration to allow the creation of an instance and the storage of it in the TeltonikaConfiguration class

TeltonikaConfiguration now is a singleton, is constructed by a factory, and is maintained using a loosely-named observer pattern. My OOP professors would be proud. Maybe a bit over-complicated, but I think it's easily digestable. Excuse my extensive use of exceptions.