Custom Asset Type Implementation *BREAKING*
Developers can now define their own asset types by extending on VehicleAsset. In TeltonikaMQTTHandler, it is as easy as changing the reference of `TELTONIKA_DEVICE_ASSET_CLASS` to your own custom type that extends the VehicleAsset. In this way, developers can create their own asset types with their own specific ways of parsing the parameters of the payload, without tweaking the code.

Due to a change in the Asset Type, current installations will have to move to this to use the new asset types.