MQTT Handler Tests Reenabled and Updated
Add a new test case to go through every single payload in the list, and check that the correct amount of parameters are being updated.

Future support should make a way to check each created attribute that checks if the correct value was parsed and the correct type was created.